+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ v1.4.2
 
 - Add movement in X axis in *scroll_element_into_view* method
 - Fix bugs in  the Dynamic Environment feature:
+
      - chars no utf-8 are accepted
      - no replace behave prefixes into a step